#10080 Removed duplicated delete code

+++ dotCMS/src/main/java/com/dotcms/content/elasticsearch/business/ESIndexAPI.java
@@ -971,7 +971,7 @@ public class ESIndexAPI {
 	 *            stream with the zipped repository file
 	 * @param snapshotName
 	 *            index to be restored
-	 * @param deleteRepository
+	 * @param cleanRepository
 	 * 	          defines if the respository should be deleted after the restore.
 	 *
 	 * @return true if the snapshot was restored
@@ -984,10 +984,9 @@ public class ESIndexAPI {
 	 * @throws IOException
 	 *             for problems writing the temporal zip file or the temporal zip contents
 	 */
-	public boolean uploadSnapshot(InputStream inputFile, boolean deleteRepository)
+	public boolean uploadSnapshot(InputStream inputFile, boolean cleanRepository)
 			throws InterruptedException, ExecutionException, ZipException, IOException {
 		File outFile = null;
-		try{
 		AdminLogger.log(this.getClass(), "uploadSnapshot", "Trying to restore snapshot index");
 		// creates specific backup path (if it shouldn't exist)
 		File toDirectory = new File(
@@ -1000,17 +999,8 @@ public class ESIndexAPI {
 		//File outFile = new File(toDirectory.getParent() + File.separator + snapshotName);
 		FileUtils.copyStreamToFile(outFile, inputFile, null);
 		ZipFile zipIn = new ZipFile(outFile);
-		boolean response = uploadSnapshot(zipIn, toDirectory.getAbsolutePath());		
-		return response;
-		}finally{
-			if(deleteRepository){
-				deleteRepository(BACKUP_REPOSITORY);
-				if(outFile!=null && outFile.exists()){
-					outFile.delete();
-				}
-			}
-		}
-		
+		boolean response = uploadSnapshot(zipIn, toDirectory.getAbsolutePath(), cleanRepository);		
+		return response;		
 	}
 
 	/**
@@ -1033,7 +1023,7 @@ public class ESIndexAPI {
 	 * @throws IOException
 	 *             for problems writing the temporal zip file or the temporal zip contents
 	 */
-	public boolean uploadSnapshot(ZipFile zip, String toDirectory)
+	public boolean uploadSnapshot(ZipFile zip, String toDirectory, boolean cleanRepository)
 			throws InterruptedException, ExecutionException, ZipException, IOException {
 		ZipUtil.extract(zip, new File(toDirectory));
 		File zipDirectory = null;
@@ -1050,10 +1040,12 @@ public class ESIndexAPI {
 			}
 			return restoreSnapshot(BACKUP_REPOSITORY, snapshotName);
 		}finally{
-			deleteRepository(BACKUP_REPOSITORY);
 			File tempZip = new File(zip.getName());
 			if(zip!=null && tempZip.exists()){
 				tempZip.delete();
+			}
+			if(cleanRepository){
+				deleteRepository(BACKUP_REPOSITORY);
 			}
 		}
 	}
